requirements.txt
openai
anthropic
whisper
transformers
torch
torchaudio
langchain
langgraph
langdetect
pydantic
azure-cognitiveservices-speech